+++ src/buttons/panel_hub_tempvoice_refresh.js
@@ -3,7 +3,7 @@ const {
 } = require("../utils/panelHubTempVoice");
 
 module.exports = {
-  customId: "panel_hub_tempvoice",
+  customId: "panel_hub_tempvoice_refresh",
 
   async execute(interaction) {
     const message = await createTempVoiceCentralMessage(interaction.guild);